% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_overlap.R
\name{CI_overlap}
\alias{CI_overlap}
\title{Confidence Measure Overlap}
\usage{
CI_overlap(lw1, up1, lw2, up2)
}
\arguments{
\item{lw1}{is the lower bound of the first confidence interval (vector or single value)}

\item{up1}{is the upper bound of the first confidence interval (vector or single value).
\code{up1} must have the same length as \code{lw1}.}

\item{lw2}{is the lower bound of the comparison confidence interval (vector or single value)}

\item{up2}{is the upper bound of the comparison confidence interval (vector or single value).
\code{up2} must have the same length as \code{lw2}.}
}
\value{
the measure of confidence interval overlap.
}
\description{
Confidence Measure Overlap
}
\details{
The length of \code{lw1} or \code{lw2} must be a multiple of the other.
}
