% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ITTtable.R
\name{compare_ITTdata}
\alias{compare_ITTdata}
\title{Dataframe to Compare ITT Across Synthetic and Confidential Data}
\usage{
compare_ITTdata(
  data.list,
  reg.models,
  response.vars = NULL,
  families = "gaussian",
  treat.vars,
  control.var = "control",
  bonferroni.npvals = NULL,
  mult.test.correct = c("treatments", "responses"),
  model.names = NULL,
  stderr.func = NULL,
  only.control.mean = FALSE,
  include.ci.overlap = FALSE,
  ci.overlap.ref.name = NULL,
  ci.confidence = 0.95,
  pivot.xtreat = FALSE,
  col.var.names = NULL,
  treat.names = NULL,
  ...
)
}
\arguments{
\item{data.list}{a named list of data sets to compare the ITT tables of. If no names are
given data sets will be named Data 1, Data 2, ...}

\item{reg.models}{vector of strings or formulas for the regression models. If
\code{reg.models==NULL} (default), then \code{treat.vars} and \code{covariate.vars},
cannot be \code{NULL} and are used to make a linear regression model for each
\code{response.vars}. If \code{reg.models} is only one formula or string, then
the predictors of that formula are used with each \code{response.vars} to make
a vector of models.}

\item{response.vars}{a vector of names, indicators, or indices of the response
variables within \code{data}.}

\item{families}{a vector of the family for each \code{reg.models} \link{\code{glm}}. The
\code{length(families)} must be 1 or equal to \code{length(reg.models)} (or \code{response.vars}).
If the length is 1, then the same distribution family is used for all models. Default is
"gaussian".}

\item{treat.vars}{a vector of variable names, indices, or logicals for the
columns of the data that are treatment variables}

\item{control.var}{the string for the name of the control variable in the data.
If \code{control.var==NULL} (default), then the control mean will not be included in the table}

\item{bonferroni.npvals}{is a numeric value for the number of p-values to use
a Bonferroni correction to adjust for multiple testing. The default is
the number of treatment variables and level coefficients in the model}

\item{mult.test.correct}{a vector with elements "treatments" or "responses". If
\code{bonferroni.npvals==NULL}, then \code{mult.test.correct} is used to determine
the bonferroni correction. If \code{mult.test.correct=="treatments"}, then
\code{bonferroni.npvals} is the number of treatment variables/levels. If
\code{mult.test.correct=="responses"}, then \code{bonferroni.npvals} is the
length of \code{reg.models} or \code{response.vars}. If \code{mult.test.correct}
is the vector of "treatments" and "responses", then \code{bonferroni.npvals} is
the number of treatment variables/levels plus the number of treatment variables/levels.}

\item{model.names}{Names for each regression model as they should appear in the table.
If NULL (default), then the variable names for the response of each
\code{reg.models} wil be used instead.}

\item{stderr.func}{(optional) is a function that takes the \link{glm} model output
and returns the standard error of the estimated treatment coefficients. If NULL
then the standard errors from the summary table will be used.}

\item{only.control.mean}{is indicator if the wider pivot data.frame only includes
the StdErr and AdjPvalue columns.}

\item{include.ci.overlap}{indicator if the overlap of the ITT confidence intervals
should be included in the outputted dataset.}

\item{ci.overlap.ref.name}{is the name of the dataset in the \code{data.list} that
should be used as the reference for all the other datasets to be compared to.
This must not be \code{NULL} if \code{include.ci.overlap==TRUE}.}

\item{ci.confidence}{is the confidence level (as a decimal) of the intervals. Default is 0.95.}

\item{pivot.xtreat}{if \code{TRUE}, then the resulting data.frame is pivoted wider
(see \link{\code{pivot_wider}}) to have "Estimate", "StdErr", "AdjPvalue" columns for
each treatment variable/level (control.var if applicable). If \code{control.var!=NULL}
and \code{only.control.mean==TRUE}, then the control mean (not the StdErr or AdjPvalue)
is also a column. If \code{pivot.xtreat==FALSE},then data frame is left in long form
with a "Treatment" column.}

\item{...}{Further arguments will be passed to the \link{kable} function.}
}
\value{
a data.frame with columns: name of dataset the model is fit on; ITT;
Std. Err; and p-value (or Adj. p-value); number of observations; treatment name;
response variable/regression model; ITT Confidence Interval Lower and Upper Bounds;
and Confidence interval overlap with a reference dataset. There is a row for
each treatment level and control for each response variable/regression model
and dataset.
}
\description{
Dataframe to Compare ITT Across Synthetic and Confidential Data
}
