% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_coef_stats.R
\name{dp_coef_stats}
\alias{dp_coef_stats}
\title{Approximate DP Estimate, Std.Error, z value, and Pr(>|z|) for a Model Coefficient}
\usage{
dp_coef_stats(
  effect.vec,
  san.sd = NA,
  bounds.sd = c(2^(-15), 2^15),
  bound.mean,
  alphas = 0.05,
  epsilon,
  delta = 0,
  pvalue.partitions = NA,
  include.san.pvalue = TRUE,
  include.ci = TRUE
)
}
\arguments{
\item{effect.vec}{vector of coefficients from the iterated fitted model}

\item{san.sd}{(optional) a sanitized standard error for use in getting the
sanitized range and then sanitized mean. If NULL, then \link{dp_estimate_sd}
is used instead.}

\item{bounds.sd}{a vector of two elements: the minimum and the maximum the standard deviation of the population \code{x} is sampled from}

\item{bound.mean}{finite values such that the mean of the population \code{x}
is sampled from is bounded by (-\code{bound.mean}, \code{bound.mean})}

\item{alphas}{a single value between (0,1) to be divided evenly by into 3 or
three values whose sum is between (0,1) such that the confidence interval is
\code{sum(alphas)}-level confidence interval. The first value is used in
the \eqn{1-\alpha/2} quantile of a standard normal. The second is used in
\eqn{b_1 log(1/\alpha_1)} where \eqn{b_1} is the scale of the Laplace noise
for the point estimate. The third is the \code{range.prob} for the
\code{dp_range()} function.}

\item{epsilon}{is the positive privacy parameter}

\item{delta}{is a non-negative privacy parameter. Default is \code{delta=0}.}

\item{pvalue.partitions}{is the number of subsets of \code{x} to use to produce p-values.
If \code{pvalue.partitions} is NA, then \code{sub.sizes} cannot be NULL and
\code{pvalue.partitions} will be set to \code{length(sub.sizes)}}

\item{include.san.pvalue}{logical to indicate if z value and P(>|z|) are returned}

\item{include.ci}{logical to indicate if sanitized confidence interval is returned}
}
\value{
if  \code{bound.mean!=NA} the vector of sanitized statistics, which include:
the point estimate; estimated standard error; lower and upper confidence
interval bounds and the confidence level (if \code{include.ci==TRUE});
the z-value and p-value (if \code{include.san.pvalue==TRUE}); and the epsilon
and delta privacy parameters. Otherwise, a vector of NA is returned.
}
\description{
Approximate DP Estimate, Std.Error, z value, and Pr(>|z|) for a Model Coefficient
}
\seealso{
\code{\link[=dp_confidence_interval(), dp_estimate_sd(), dp_pvalue]{dp_confidence_interval(), dp_estimate_sd(), dp_pvalue()}}
}
