% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_karwa_vadan_algorithms.R
\name{dp_confidence_interval}
\alias{dp_confidence_interval}
\title{Confidence Intervals satisfying Approximate-DP}
\usage{
dp_confidence_interval(
  x,
  epsilon.vec,
  delta.vec = 0,
  alphas = 0.05,
  san.point = NA,
  bounds.sd = c(2^(-15), 2^15),
  x.sd = NA,
  bound.mean = NA,
  return.point.sd = FALSE
)
}
\arguments{
\item{x}{vector of continuous numeric confidential values.}

\item{epsilon.vec}{a positive value or a vector of them for the privacy parameters. (See details).}

\item{delta.vec}{a nonnegative value or a vector of them for the privacy parameters. Default is 0. (See details).}

\item{alphas}{a single value between (0,1) to be divided evenly by into 3 or
three values whose sum is between (0,1) such that the confidence interval is
\code{sum(alphas)}-level confidence interval. The first value is used in
the \eqn{1-\alpha/2} quantile of a standard normal. The second is used in
\eqn{b_1 log(1/\alpha_1)} where \eqn{b_1} is the scale of the Laplace noise
for the point estimate. The third is the \code{range.prob} for the
\code{dp_range()} function.}

\item{san.point}{(optional) the sanitized point estimate of the coefficient.
If \code{san.point=NA} (default), then \code{\link{dp_range}} will be used
with the \code{\link{laplace_mechanism}} to get a sanitized point estimate.}

\item{bounds.sd}{a vector of two elements: the minimum and the maximum the standard deviation of the population \code{x} is sampled from}

\item{x.sd}{(optional) the publicly known (or privacy-preserving estimate of)
standard deviation of the population \code{x} is sampled from. If it is NA,
then the \code{dp_estimate_sd} function will be used to get a DP estimate of it.
In this case, \code{bounds.sd} must be a tuple of the bounds on the
standard deviation. (See details for specifications of \code{epsilon.vec}
and \code{delta.vec} when \code{x.sd=NA})}

\item{bound.mean}{finite values such that the mean of the population \code{x}
is sampled from is bounded by (-\code{bound.mean}, \code{bound.mean})}

\item{return.point.sd}{logical indicator for whether the privacy-preserving point
and standard deviation estimate of treatment effect should be returned.}
}
\value{
a vector of the lower and upper bounds of the
\code{sum(alphas)}-level confidence interval that satisfies
(\code{sum(epsilon.vec)}, \code{sum(delta.vec)})-DP, or a list of the
privacy-preserving confidence interval with the privacy-preserving point
and standard deviation estimates for the treatment.
}
\description{
Algorithm 4 by Karwa and Vadhan, (2017)
}
\details{
For the \code{epsilon.vec} and \code{delta.vec}, If the \code{x.sd=NA} is
provided, \code{epsilon.vec} should be 3 values (or 1 value to be split evenly
across 3 privacy steps) and the \code{delta.vec} should be 2 values (or 1
value to be split evenly). Otherwise, \code{epsilon.vec} should be 2 values
(or 1 to be split evenly) and \code{delta.vec} should be 1.
}
\examples{
set.seed(1)



}
\references{
Karwa, V. and Vadhan, S. (2017). Finite sample differentially private confidence intervals.
}
\seealso{
Other ApproximateDPmeasures: 
\code{\link{dp_estimate_sd}()},
\code{\link{dp_pvalue}()},
\code{\link{dp_range}()}
}
\concept{ApproximateDPmeasures}
