% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_karwa_vadan_algorithms.R
\name{dp_estimate_sd}
\alias{dp_estimate_sd}
\title{Sanitized estimate of standard deviation satisfying Approximate-DP}
\usage{
dp_estimate_sd(x, epsilon, delta = 0, bounds.sd)
}
\arguments{
\item{x}{vector of continuous numeric confidential values.}

\item{epsilon}{a positive privacy parameter}

\item{delta}{a nonnegative privacy parameter. Defaul is \code{delta=0}}

\item{bounds.sd}{a vector of two elements: the minimum and the maximum the standard deviation of the population \code{x} is sampled from}
}
\value{
a sanitized estimated variance of \code{x}
}
\description{
Algorithm 2 by Karwa and Vadhan (2017)
}
\examples{
set.seed(1)
x=rnorm(500,70,5)
dp_estimate_sd(x=x,epsilon=1,delta=0.001,bounds.sd=c(2^(-15),2^15))


}
\references{
{
Karwa, V. and Vadhan, S. (2017). Finite sample differentially private confidence intervals.

}
}
\seealso{
Other ApproximateDPmeasures: 
\code{\link{dp_confidence_interval}()},
\code{\link{dp_pvalue}()},
\code{\link{dp_range}()}
}
\concept{ApproximateDPmeasures}
