% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_model_summary.R
\name{dp_model_summary}
\alias{dp_model_summary}
\title{Approximate DP Model Summary Table}
\usage{
dp_model_summary(
  formula,
  confidential.data,
  epsilon.list,
  delta.list = 0,
  bd.sd.list = c(2^(-15), 2^(15)),
  bd.mean.list,
  alphas.list = 0.05,
  p.partitions.list,
  num.iters = 10000,
  treatment.var = NULL,
  rseed = NA,
  return.time = FALSE,
  just.treatment = FALSE,
  return.confidential.table = FALSE,
  use.san.residerror = TRUE,
  synth.data = NULL,
  synth.epsilon = NA,
  synth.delta = 0,
  continuous.vars = NULL,
  num.bin = NULL,
  bin.param = NULL,
  add.cont.variation = TRUE,
  assign.type = "simple",
  blocks = NULL,
  clusters = NULL,
  within.blocks = TRUE,
  ...
)
}
\arguments{
\item{formula}{either string or \code{\link{as.formula}} object
with response variable as the first variable in the formula.
If \code{treatment.var} is not supplied, the second variable in the formula
is taken as the treatment variable.}

\item{confidential.data}{a data.frame object with the confidential data to be used.}

\item{epsilon.list}{a list where each entry is the epsilon value(s) for the
corresponding model coefficient}

\item{delta.list}{a list where each entry is the delta value(s) for the
corresponding model coefficient}

\item{bd.sd.list}{a list where each entry is the bounds.sd tuple for the
corresponding model coefficient (see \code{\link{dp_confidence_interval}})}

\item{bd.mean.list}{a list where each entry is the bound.mean value for the
corresponding model coefficient (see \code{\link{dp_confidence_interval}})}

\item{alphas.list}{a list where each entry is the alpha value(s) for the
corresponding model coefficient (see the first 1 (or 3) values are passed to
\code{\link{dp_confidence_interval}} and the last value is passed to
\code{\link{dp_pvalue}})}

\item{p.partitions.list}{a list where each entry is the pvalue.partitions
value for the corresponding model coefficient. (see \code{\link{dp_pvalue}})}

\item{num.iters}{the number of proxy response variable to produce}

\item{treatment.var}{a vector of column indices, logical, or strings of column
names to denote what columns of \code{confidential.data} are treatment columns}

\item{rseed}{a value to set the randomization seed. If \code{rseed=NA} (default),
then no random seed it set.}

\item{return.time}{indicator if computation time should be returned}

\item{just.treatment}{indicator if just the treatment variable sanitized
statistics (rather than all predictors) should be returned}

\item{return.confidential.table}{indicator if the confidential model summary
table should be returned}

\item{use.san.residerror}{indicator if a sanitized residual error should be used.}

\item{synth.data}{a data.frame with the covariates including any blocking or
clustering variables to be used in assignment}

\item{synth.epsilon}{is the privacy parameter for making synthetic treatment and
covariate data. If no \code{synth.data} is supplied and \code{use.san.residerr==TRUE},
then synth.epsilon must be a positive value.}

\item{synth.delta}{is the privacy parameter for making synthetic treatment and
covariate data. Default is 0.}

\item{continuous.vars}{vector of strings, indices, or logical values to indicate
the columns of \code{data} with continuous variables. If \code{continuous.vars=NULL},
all variables are assumed to be categorical.}

\item{num.bin}{a numeric value or vector of numeric values for the number of
bins for each continuous variable. If \code{num.bin=NA}, the \code{bin.param}
is used to determine the number of bins.}

\item{bin.param}{a numeric value in (0,1) to determine number of bins for
continuous variables to be used if \code{num.bins=NA}.}

\item{add.cont.variation}{a logical if \code{TRUE}, add uniform noise to midpoint
to get a random value within the range of bin. Default is \code{FALSE}.}

\item{assign.type}{a string taking values "simple","complete","block",
"cluster", or "block_and_cluster" to specify what kind of random assignment
should be used. Default is "simple".}

\item{blocks}{column name(s) for the variables for block treatment assignment. This
is only needed if \code{assign.type=="block"} or \code{=="block_and_cluster"}.}

\item{clusters}{is a vector of indices, column names, or logical to specify
what columns of \code{synth.data} to use as clustering variables in random
treatment assignment. If \code{assign.treatment} takes value "cluster" or
"block_and_cluster" a value for \code{clusters} must be supplied.}

\item{within.blocks}{a logical to indicate if the data should be sampled within the
block levels in order to maintain the same block sizes as the original data.}

\item{...}{additional arguments are such as \code{blocks} or \code{clusters}
are supplied to the \code{treatment_assign} function.}
}
\value{
a table with sanitized point estimate, standard deviation estimate,
lower and upper 95\\% confidence interval bounds, z value and P(>|z|)
for each model coefficient.
}
\description{
Based on Algorithm 3 from paper.
}
\details{
If only one value is given for \code{epsilon.list} or \code{delta.list}, then
this is taken to be the budget for the whole table and is divided evenly across the
For the elements of \code{returntypes}:
If "synth.data" is included, the full synthetic dataset is returned.
If "san.treat.effect", the sanitized (or privacy-preserving) treatment effect from fitting a model with the synthetic data is returned.
If "san.model" is included, the \code{\link{glm}} object of the model fitted with the synthetic data is outputted.
If "confidential.model" is included, the \code{glm} object of the model fitted with the confidential data is outputted without privacy protections.
If "comp.time" is included, the computation time of the algorithm is outputted. If "privacy.cost" is included, the overall privacy cost is returned as a vector of c(composed epsilon,composed delta).
Additional parameters ... are supplied to the \code{glm} function.
To specify random treatment assignment further. Use \code{\link{synthdata_perturb_mvhist}} with \code{with.treatment=TRUE} to create your synthetic data first.
Users may bring their own synthetic data and use \code{\link{treatment_assign}} to further control the specifications of the random treatment assignment methodology.
Then the synthetic dataset can be supplied to this function as \code{synth.data} input.
}
