% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_perturbed_hist.R
\name{dp_perturbed_hist}
\alias{dp_perturbed_hist}
\title{Generate sanitized proportions that satisfy (epsilon,delta)-DP}
\usage{
dp_perturbed_hist(hist.df, epsilon, delta = 0)
}
\arguments{
\item{hist.df}{a data.frame of the counts for the number of occurrences of a category (or combination of categories) which is column \code{hist.df$Freq}vector of continuous values to be split into bins}

\item{epsilon}{a positive privacy parameter value}

\item{delta}{a non-negative value for a privacy parameter}
}
\value{
a data.frame which is the hist.df with column \code{hist.df$san.prop} which is a sanitized proportion of the data that is in a bin which is generated by a Laplace mechanism satisfying (\code{epsilon},\code{delta})-differential privacy.
}
\description{
Generate sanitized proportions that satisfy (epsilon,delta)-DP
}
\examples{
freq.df=multivariate_histogram(mtcars[,2:4],continuous.vars=c("disp","hp"),num.bin=3)
dp_perturbed_hist(freq.df,0.9,delta=0.01)

}
\references{
When \code{delta==0}, perturbed histogram algorithm is from:
Karwa, V. and Vadhan, S. (2017). Finite sample differentially private confidence intervals.

When \code{delta>0}, perturbed histogram algorithm is from:
Bun, M., Nissim, K., and Stemmer, U. (2016). Simultaneous private learning of multiple
concepts. In Proceedings of the 2016 ACM Conference on Innovations in Theoretical Computer Science,
ITCS ’16, page 369–380, New York, NY, USA. Association for Computing Machinery.
}
\seealso{
Other syntheticData: 
\code{\link{multivariate_histogram}()},
\code{\link{synthdata_perturb_mvhist}()},
\code{\link{treatment_assign}()}
}
\concept{syntheticData}
