% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pvalue.R
\name{dp_pvalue}
\alias{dp_pvalue}
\title{P-Values from Approximate DP}
\usage{
dp_pvalue(
  x,
  test,
  epsilon,
  delta = 0,
  alpha = 0.05,
  pvalue.partitions = NA,
  sub.sizes = NULL,
  ...
)
}
\arguments{
\item{x}{is confidential data to be tested}

\item{test}{is a function that tests data \code{x} and returns a p-value}

\item{epsilon}{is the positive privacy parameter}

\item{delta}{is a non-negative privacy parameter. Default is \code{delta=0}.}

\item{alpha}{is the significance level of the tests. Default is 0.05}

\item{pvalue.partitions}{is the number of subsets of \code{x} to use to produce p-values.
If \code{pvalue.partitions} is NA, then \code{sub.sizes} cannot be NULL and
\code{pvalue.partitions} will be set to \code{length(sub.sizes)}}

\item{sub.sizes}{a vector of the size for each subset of the partition of \code{x}.
If \code{pvalue.partitions} is not NA and \code{sub.sizes} is not NULL,
\code{pvalue.partitions} will be used to override \code{sub.sizes}.}

\item{...}{any additional parameters are inputted into the \code{test} function.}
}
\value{
a list of the sanitized p-value and the test statistic \code{z}.
}
\description{
Algorithm 1 by Kazan et al.(2023) and uses functions from binomialDP package (Tran,2020).
}
\examples{
set.seed(1)
test.func=function(x){pnorm(mean(x),lower.tail=FALSE)}
dp_pvalue(x=rnorm(500,1.2,1),epsilon=1,test=test.func,alpha=0.05,pvalue.partitions=12)

}
\references{
\code{dp_pvalue} is Algorithm 1 in
Kazan, Z., Shi, K., Groce, A.,and Bray, A. P. (2023). The test of tests:
A framework for differentially private hypothesis testing. In Krause, A.,
Brunskill, E., Cho, K., Engelhardt, B., Sabato, S.,and Scarlett, J., editors,
Proceedings of the 40th International Conference on Machine Learning,
volume 202 of Proceedings of Machine Learning Research, pages 16131–16151. PMLR.

\if{html}{\out{<div class="sourceCode">}}\preformatted{\code{dp_pvalue} used binomialDP package from GitHub.
Tran, T. (2020). Differentiallly private inference for binomial data.
\url{https://github.com/tranntran/binomialDP.}

binomialDP uses results from:
Awan, J. and Slavkovi´c, A. (2018). Differentially private uniformly most
   powerful tests for binomial data. Advances in Neural Information
   Processing Systems, 31:4208–4218.
}\if{html}{\out{</div>}}
}
\seealso{
Other ApproximateDPmeasures: 
\code{\link{dp_confidence_interval}()},
\code{\link{dp_estimate_sd}()},
\code{\link{dp_range}()}
}
\concept{ApproximateDPmeasures}
