% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_karwa_vadan_algorithms.R
\name{dp_range}
\alias{dp_range}
\title{Sanitized max and minimum satisfying Approximate-DP with known variance.}
\usage{
dp_range(x, sd, epsilon, delta = 0, bound.mean, range.prob)
}
\arguments{
\item{x}{vector of continuous numeric confidential values.}

\item{sd}{a public standard deviation of the population \code{x} is sampled from.
A sanitized estimate of it may be used (see details.)}

\item{epsilon}{a positive privacy parameter}

\item{delta}{a privacy parameter. Default is 0}

\item{bound.mean}{finite values such that the mean of the population \code{x}
is sampled from is bounded by (-\code{bound.mean}, \code{bound.mean})}

\item{range.prob}{numeric value so the probability that
\eqn{P(\bigcup_{i=1}^n X_{min}\leq X_i\leq X_{max})\geq 1-}$\code{range.prob}
where $X_i$ are i.i.d. normal random variable with mean and variance equal
to that of the population \code{x} is sampled from and $n$ is the length of \code{x}.}
}
\value{
a vector with sanitized minimum and maximum of \code{x}.
}
\description{
Algorithm 1 by Karwa and Vadhan(2017)
}
\details{
If \code{sd} is unknown or not public, the \code{dp_estimate_sd()} function
can be used to get a sanitized estimate of the standard deviation to use in this function.
}
\examples{
set.seed(1)
x=rnorm(500,70,5)
print(c(min(x),max(x)))
dp_range(x=x,sd=5,epsilon=1,delta=0.001,bound.mean=1e6,range.prob=0.025)


}
\references{
{
Karwa, V. and Vadhan, S. (2017). Finite sample differentially private confidence intervals.

}
}
\seealso{
Other ApproximateDPmeasures: 
\code{\link{dp_confidence_interval}()},
\code{\link{dp_estimate_sd}()},
\code{\link{dp_pvalue}()}
}
\concept{ApproximateDPmeasures}
