% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_synth.R
\name{dp_synthdata}
\alias{dp_synthdata}
\title{Approximate DP Synthetic Data and Treatment Effect Table}
\usage{
dp_synthdata(
  formula,
  confidential.data,
  synth.data = NULL,
  synth.epsilon = NA,
  synth.delta = 0,
  continuous.vars = NULL,
  num.bin = NULL,
  bin.param = NA,
  assign.type = "simple",
  blocks = NULL,
  clusters = NULL,
  within.blocks = TRUE,
  epsilon.list,
  delta.list = 0,
  bd.sd.list = c(2^(-15), 2^(15)),
  bd.mean.list,
  alphas.list = 0.05,
  num.iters = 10000,
  treatment.var = NULL,
  rseed = NA,
  return.time = FALSE,
  return.confidential.table = FALSE,
  return.san.summary = TRUE,
  use.san.residerror = TRUE,
  return.treatment.pvalue = TRUE,
  return.treatment.CI = TRUE,
  treat.ppart.list = NULL,
  ...
)
}
\arguments{
\item{formula}{either string or \code{\link{as.formula}} object
with response variable as the first variable in the formula.
If \code{treatment.var} is not supplied, the second variable in the formula
is taken as the treatment variable.}

\item{confidential.data}{a data.frame object with the confidential data to be used.}

\item{synth.data}{a data.frame with the covariates including any blocking or
clustering variables to be used in assignment}

\item{synth.epsilon}{privacy parameter for synthetic covariate and treatment data}

\item{synth.delta}{privacy parameter for synthetic covariate and treatment data. Default 0}

\item{continuous.vars}{vector of strings, indices, or logical values to indicate
the columns of \code{data} with continuous variables. If \code{continuous.vars=NULL},
all variables are assumed to be categorical.}

\item{num.bin}{a numeric value or vector of numeric values for the number of
bins for each continuous variable. If \code{num.bin=NA}, the \code{bin.param}
is used to determine the number of bins.}

\item{bin.param}{a numeric value in (0,1) to determine number of bins for
continuous variables to be used if \code{num.bins=NA}.}

\item{assign.type}{a string taking values "simple","complete","block",
"cluster", or "block_and_cluster" to specify what kind of random assignment
should be used. Default is "simple".}

\item{blocks}{column name(s) for the variables for block treatment assignment. This
is only needed if \code{assign.type=="block"} or \code{=="block_and_cluster"}.}

\item{clusters}{is a vector of indices, column names, or logical to specify
what columns of \code{synth.data} to use as clustering variables in random
treatment assignment. If \code{assign.treatment} takes value "cluster" or
"block_and_cluster" a value for \code{clusters} must be supplied.}

\item{within.blocks}{a logical to indicate if the data should be sampled within the
block levels in order to maintain the same block sizes as the original data.}

\item{epsilon.list}{a list where each entry is the epsilon value(s) for the
corresponding model coefficient}

\item{delta.list}{a list where each entry is the delta value(s) for the
corresponding model coefficient}

\item{bd.sd.list}{a list where each entry is the bounds.sd tuple for the
corresponding model coefficient (see \code{\link{dp_confidence_interval}})}

\item{bd.mean.list}{a list where each entry is the bound.mean value for the
corresponding model coefficient (see \code{\link{dp_confidence_interval}})}

\item{alphas.list}{a list where each entry is the alpha value(s) for the
corresponding model coefficient (see the first 1 (or 3) values are passed to
\code{\link{dp_confidence_interval}} and the last value is passed to
\code{\link{dp_pvalue}})}

\item{num.iters}{the number of proxy response variable to produce}

\item{treatment.var}{a vector of column indices, logical, or strings of column
names to denote what columns of \code{confidential.data} are treatment columns}

\item{rseed}{a value to set the randomization seed. If \code{rseed=NA} (default),
then no random seed it set.}

\item{return.time}{indicator if computation time should be returned}

\item{return.confidential.table}{indicator if the confidential model summary
table should be returned}

\item{use.san.residerror}{indicator if a sanitized residual error should be used.}

\item{return.treatment.pvalue}{indicator to return pvalues for treatment effects}

\item{return.treatment.CI}{indicator to return confidence intervals for treatment effects
(last \code{alphas.list} value(s) will be used for confidence level)}

\item{treat.ppart.list}{list of p-value partitions for each treatment effect}

\item{...}{additional arguments are such as \code{blocks} or \code{clusters}
are supplied to the \code{treatment_assign} function.}
}
\value{
synthetic data set (and possible treatment pvalue, CI and confidential summary table.)
}
\description{
Approximate DP Synthetic Data and Treatment Effect Table
}
\details{
To become the example:
tempdf=mtcars\link{,1:6}
tempdf$am=as.factor(mtcars$am)
tempdf$cyl=as.factor(mtcars$cyl)
tform="mpg~am+cyl+hp+wt"
tcont=c("hp","wt")
dp_synthdata(tform,tempdf,synth.epsilon = 1,continuous.vars = tcont,
num.bin=3,epsilon.list=1,bd.mean.list=100,num.iters=300,use.san.residerror = T,
treat.ppart.list = 5)
}
