% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mechanisms.R
\name{gaussian_mechanism}
\alias{gaussian_mechanism}
\title{Gaussian Mechanism}
\usage{
gaussian_mechanism(x, epsilon, delta, f = NULL, sensitivity = NULL, rseed = NA)
}
\arguments{
\item{x}{vector of numeric confidential values}

\item{epsilon}{a positive value for the privacy cost.}

\item{delta}{a numeric positive value for the privacy cost}

\item{f}{a function applied to \code{x} which will have its result protected
with Gaussian noise. If \code{f=NULL}, then it is the identity function (i.e. f(x)=x).}

\item{sensitivity}{the L2 sensitivity of \code{f}}

\item{rseed}{a numeric value if supplied sets the random seed}
}
\value{
randomized output that is generated with Gaussian noise to satisfy
approximate differential privacy.
}
\description{
Gaussian Mechanism infuses noise to satisfy approximate differential privacy.
}
\examples{
set.seed(1)
x=rbinom(50,20,0.5)/20
gaussian_mechanism(x,epsilon=1,delta=0.001,sensitivity=1)


}
\seealso{
Other DPmechanisms: 
\code{\link{laplace_mechanism}()},
\code{\link{randomized_response}()}
}
\concept{DPmechanisms}
