% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_dp.R
\name{hybrid_synth}
\alias{hybrid_synth}
\title{Hybrid-DP creates a synthetic dataset for RCTs using DP-based methods}
\usage{
hybrid_synth(
  formula,
  family = "gaussian",
  confidential.data,
  epsilon = NA,
  delta = 0,
  treatment.var = NULL,
  synth.data = NULL,
  continuous.vars = NULL,
  num.bin = NULL,
  bin.param = NA,
  add.cont.variation = FALSE,
  assign.type = "simple",
  blocks = NULL,
  within.blocks = TRUE,
  clusters = NULL,
  returntypes = c("synth.data", "san.model"),
  rseed = NA,
  ...
)
}
\arguments{
\item{formula}{either string or \code{\link{as.formula}} object
with response variable as the first variable in the formula.
If \code{treatment.var} is not supplied, the second variable in the formula
is taken as the treatment variable.}

\item{family}{a description of the error distribution and link function to be
supplied to \code{\link{glm}}.}

\item{confidential.data}{a data.frame object with the confidential data to be used.}

\item{epsilon}{a positive value that defines the privacy cost}

\item{delta}{a nonnegative value that defines the privacy cost. Default is \code{delta=0}.}

\item{treatment.var}{a vector of column indices, logical, or strings of column
names to denote what columns of \code{confidential.data} are treatment columns}

\item{synth.data}{a data.frame with the covariates including any blocking or
clustering variables to be used in assignment}

\item{continuous.vars}{vector of strings, indices, or logical values to indicate
the columns of \code{data} with continuous variables. If \code{continuous.vars=NULL},
all variables are assumed to be categorical.}

\item{num.bin}{a numeric value or vector of numeric values for the number of
bins for each continuous variable. If \code{num.bin=NA}, the \code{bin.param}
is used to determine the number of bins.}

\item{bin.param}{a numeric value in (0,1) to determine number of bins for
continuous variables to be used if \code{num.bins=NA}.}

\item{add.cont.variation}{a logical if \code{TRUE}, add uniform noise to midpoint
to get a random value within the range of bin. Default is \code{FALSE}.}

\item{assign.type}{a string taking values "simple","complete","block",
"cluster", or "block_and_cluster" to specify what kind of random assignment
should be used. Default is "simple".}

\item{blocks}{column name(s) for the variables for block treatment assignment. This
is only needed if \code{assign.type=="block"} or \code{=="block_and_cluster"}.}

\item{within.blocks}{a logical to indicate if the data should be sampled within the
block levels in order to maintain the same block sizes as the original data.}

\item{clusters}{is a vector of indices, column names, or logical to specify
what columns of \code{synth.data} to use as clustering variables in random
treatment assignment. If \code{assign.treatment} takes value "cluster" or
"block_and_cluster" a value for \code{clusters} must be supplied.}

\item{returntypes}{a vector of strings where the elements can take values
\code{"synth.data"}, \code{"san.treat.effect"}, \code{"san.model"},
and/or \code{"confidential.model"}. These determine what values are returned
from the function.}

\item{rseed}{a value to set the randomization seed. If \code{rseed=NA} (default),
then no random seed it set.}

\item{...}{
  Arguments passed on to \code{\link[stats:glm]{stats::glm}}
  \describe{
    \item{\code{data}}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{glm} is called.}
    \item{\code{weights}}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.}
    \item{\code{subset}}{an optional vector specifying a subset of observations
    to be used in the fitting process.}
    \item{\code{na.action}}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is set by
    the \code{na.action} setting of \code{\link{options}}, and is
    \code{\link[stats]{na.fail}} if that is unset.  The \sQuote{factory-fresh}
    default is \code{\link[stats]{na.omit}}.  Another possible value is
    \code{NULL}, no action.  Value \code{\link[stats]{na.exclude}} can be useful.}
    \item{\code{start}}{starting values for the parameters in the linear predictor.}
    \item{\code{etastart}}{starting values for the linear predictor.}
    \item{\code{mustart}}{starting values for the vector of means.}
    \item{\code{offset}}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one is
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}
    \item{\code{control}}{a list of parameters for controlling the fitting
    process.  For \code{glm.fit} this is passed to
    \code{\link[stats]{glm.control}}.}
    \item{\code{model}}{a logical value indicating whether \emph{model frame}
    should be included as a component of the returned value.}
    \item{\code{method}}{the method to be used in fitting the model.  The default
    method \code{"glm.fit"} uses iteratively reweighted least squares
    (IWLS): the alternative \code{"model.frame"} returns the model frame
    and does no fitting.

    User-supplied fitting functions can be supplied either as a function
    or a character string naming a function, with a function which takes
    the same arguments as \code{glm.fit}.  If specified as a character
    string it is looked up from within the \pkg{stats} namespace.
  }
    \item{\code{x,y}}{For \code{glm}:
    logical values indicating whether the response vector and model
    matrix used in the fitting process should be returned as components
    of the returned value.

    For \code{glm.fit}: \code{x} is a design matrix of dimension
    \code{n * p}, and \code{y} is a vector of observations of length
    \code{n}.
  }
    \item{\code{singular.ok}}{logical; if \code{FALSE} a singular fit is an
    error.}
    \item{\code{contrasts}}{an optional list. See the \code{contrasts.arg}
    of \code{model.matrix.default}.}
  }}
}
\value{
a list or element determined by \code{returnvalues}. The elements are
returned in a list in the same order as the vector requests them. (See details)
}
\description{
(Algorithm 1 in paper) which uses estimates of coefficients,
treatment effect, and standard error from the confidential data to create
a synthetic dataset and estimated treatment effect based on differential
privacy methods.
}
\details{
For the elements of \code{returntypes}:
If "synth.data" is included, the full synthetic dataset is returned.
If "san.treat.effect", the sanitized (or privacy-preserving) treatment effect
from fitting a model with the synthetic data is returned.
If "san.model" is included, the \code{\link{glm}} object
of the model fitted with the synthetic data is outputted.
If "confidential.model" is included, the \code{\link{glm}} object
of the model fitted with the confidential data is outputted without privacy protections.
If "comp.time" is included, the computation time of the algorithm is outputted.

Additional parameters are passed to the \code{\link{glm}} function.
To specify random treatment assignment further. Use \code{synthdata_perturb_mvhist()}
with \code{with.treatment=TRUE} to create your synthetic data first.
Users may bring their own synthetic data and use \code{treatment_assign} to
further control the specifications of the random treatment assignment methodology.
Then the synthetic dataset can be supplied to this function as \code{synth.data} input.
}
