% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ITTtable.R
\name{ITTtable_oneresponse}
\alias{ITTtable_oneresponse}
\title{Get ITT table values for a given regression model.}
\usage{
ITTtable_oneresponse(
  data,
  reg.model,
  family = "gaussian",
  treat.vars,
  bonferroni.npvals = NULL,
  control.var = NULL,
  add.pval.stars = TRUE,
  stderr.func = NULL,
  incl.df = FALSE,
  ...
)
}
\arguments{
\item{data}{is the data to use to fit the model}

\item{reg.model}{is the regression formula to fit}

\item{family}{is the family for the glm to use to fit the model. Default is "gaussian".}

\item{treat.vars}{a vector of variable names, indices, or logicals for the
columns of the data that are treatment variables}

\item{bonferroni.npvals}{is a numeric value for the number of p-values to use
a Bonferroni correction to adjust for multiple testing. The default is
the number of treatment variables and level coefficients in the model}

\item{control.var}{the string for the name of the control variable in the data.
If \code{control.var==NULL} (default), then the control mean will not be included in the table}

\item{add.pval.stars}{if \code{TRUE}, then standard errors will be formatted with
3 decimal places in square brackets followed by one "*" if unadjusted p-value is <0.1,
two if <0.05, and three if <0.01. For example, \code{"[0.110]**"}. If
\code{add.pval.stars==FALSE}, then standard error will remain as numeric elements.}

\item{stderr.func}{(optional) is a function that takes the \link{glm} model output
and returns the standard error of the estimated treatment coefficients. If NULL
then the standard errors from the summary table will be used.}

\item{...}{additional inputs will be passed to \code{\link{glm}}}
}
\value{
data.frame with columns for the estimates, standard error (with stars
for unadjusted p-value if \code{add.pval.stars==TRUE}),  Bonferroni corrected p-values,
treatment variable/level names (or control), and response variable name.
With rows for each treatment variable/level and level and the control variable
(if \code{control.var!=NULL}).
}
\description{
Get ITT table values for a given regression model.
}
