% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mechanisms.R
\name{laplace_mechanism}
\alias{laplace_mechanism}
\title{Laplace Mechanism}
\usage{
laplace_mechanism(x, epsilon, f = NULL, sensitivity = NULL, rseed = NA)
}
\arguments{
\item{x}{vector of numeric confidential values}

\item{epsilon}{a positive value for the privacy cost.}

\item{f}{a function applied to \code{x} which will have its result protected
with Gaussian noise. If \code{f=NULL}, then it is the identity function (i.e. f(x)=x).}

\item{sensitivity}{the L2 sensitivity of \code{f}}

\item{rseed}{a numeric value if supplied sets the random seed}
}
\value{
randomized output that is generated with Laplace noise to satisfy
pure differential privacy.
}
\description{
Laplace Mechanism infuses noise to satisfy pure differential privacy.
}
\examples{
set.seed(1)
x=rbinom(50,20,0.5)/20
laplace_mechanism(mtcars$mpg,epsilon=1,sensitivity=1)


}
\seealso{
Other DPmechanisms: 
\code{\link{gaussian_mechanism}()},
\code{\link{randomized_response}()}
}
\concept{DPmechanisms}
