% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_histogram.R
\name{multivariate_histogram}
\alias{multivariate_histogram}
\title{Generate a frequency data.frame for multivariate histogram of mixed variable types}
\usage{
multivariate_histogram(
  data,
  continuous.vars = NULL,
  num.bin = NULL,
  bin.param = NA,
  which.cont.out = FALSE,
  return.time = TRUE
)
}
\arguments{
\item{data}{a data.frame to generate (epsilon,delta) differentially private version of}

\item{continuous.vars}{list of strings for the names, vector of indexes, or vector
of logical for columns of data with continuous variables}

\item{num.bin}{a numeric value or vector of numeric values for the number of bins
for each continuous variable. If =NA use bin.param to determine the number of bins}

\item{bin.param}{a numeric value in (0,1) to determine number of bins for continuous
variables to be used if num.bins==NA.}

\item{which.cont.out}{a logical if TRUE, return which.count and histogram data.frame.
Default FALSE only returns histogram data.frame}

\item{return.time}{a logical to indicate if the computation times should be
returned with the synthetic data. Default is \code{FALSE}.}
}
\value{
if which.count.out=FALSE, returns a data frame with frequencies of each combination of categorical and discretized continuous variables. If which.cont.out=TRUE and there are continuous variables, returns a list with the first element is the mv histogram and the second is the which.cont variable.
}
\description{
Generate a frequency data.frame for multivariate histogram of mixed variable types
}
\examples{
multivariate_histogram(mtcars[,2:4],continuous.vars=c("disp","hp"),num.bin=3)

}
\seealso{
Other syntheticData: 
\code{\link{dp_perturbed_hist}()},
\code{\link{synthdata_perturb_mvhist}()},
\code{\link{treatment_assign}()}
}
\concept{syntheticData}
