% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_returntypes.R
\name{output_returntypes}
\alias{output_returntypes}
\title{Internal-use function to parse the returntype variable for the three algorithm code functions.}
\usage{
output_returntypes(
  returntypes,
  synth.data = NULL,
  san.treatment.effect = NA,
  san.model = NULL,
  treatment.var = NULL,
  confidential.model = NULL,
  start.time = NULL,
  time.values = NULL,
  synth.priv.cost = NULL,
  ...
)
}
\arguments{
\item{returntypes}{a vector of strings where the elements can take values
\code{"synth.data"}, \code{"san.treat.effect"}, \code{"san.model"},
and/or \code{"confidential.model"}. These determine what values are returned
from the function.}

\item{synth.data}{a data.frame with the covariates including any blocking or
clustering variables to be used in assignment}

\item{treatment.var}{a vector of column indices, logical, or strings of column
names to denote what columns of \code{confidential.data} are treatment columns}

\item{synth.priv.cost}{is a named vector with the epsilon and delta totals
for the generation of the synthetic data}

\item{...}{additional arguments are such as \code{blocks} or \code{clusters}
are supplied to the \code{treatment_assign} function.}

\item{treat.stat.epsilon}{a vector or single value for the epsilon used for the
mechanisms to return the privacy preserving statistics about the treatment effect.}

\item{treat.stat.delta}{a vector or single value for the delta used for the
mechanisms to return the privacy preserving statistics about the treatment effect.}
}
\value{
list of objects to be returned from \code{hybrid_dp}, \code{crude_fully_dp}, or \code{advanced_fully_dp}
}
\description{
Internal-use function to parse the returntype variable for the three algorithm code functions.
}
\keyword{internal}
