% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mechanisms.R
\name{randomized_response}
\alias{randomized_response}
\title{Randomized Response changes a binary (0,1) variable with some probability}
\usage{
randomized_response(binary.vec, epsilon = NA, prob.flip = NA, rseed = NA)
}
\arguments{
\item{binary.vec}{vector of 0,1 values (or TRUE FALSE values)}

\item{epsilon}{a positive numeric privacy parameter. If supplied,
the randomized output satifies \code{epsilon}-differential privacy}

\item{prob.flip}{(optional) a numeric value between 0 and 1. If \code{epsilon=NA},
then \code{prob.flip} cannot be \code{NA} and it determines the probability
any binary entry is flipped.}

\item{rseed}{a numeric value if supplied sets the random seed}
}
\value{
randomized binary vector using randomized response.
}
\description{
Randomized Response changes a binary (0,1) variable with some probability
}
\examples{
set.seed(1)
x=c(rep(0,20),rep(1,20))
randomized_response(x,epsilon=1)


}
\seealso{
Other DPmechanisms: 
\code{\link{gaussian_mechanism}()},
\code{\link{laplace_mechanism}()}
}
\concept{DPmechanisms}
