% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response_glm.R
\name{simulate_response_glm}
\alias{simulate_response_glm}
\title{Simulate Responses Values Given a Model and New Data}
\usage{
simulate_response_glm(
  mod,
  newdata,
  predictor.formula = NULL,
  nsim = 1,
  cov.mat = NULL
)
}
\arguments{
\item{mod}{glm model object}

\item{newdata}{a data.frame of the new data of predictors to generate responses from}

\item{predictor.formula}{a formula object that is the formula used in the \code{mod} object.
If \code{NULL}, the formula is extracted from \code{mod}.}

\item{nsim}{a integer for the number of responses to simulate from the \code{newdata}}

\item{cov.mat}{the covariance matrix for the estimated model coefficients. If NULL
(default), then \code{stats::vcov(mod)} is used.}
}
\value{
a vector (or matrix) with \code{nsim} columns and \code{nrows(newdata)} rows
of response randomly generated from the \code{mod} using \code{newdata} as the
predictors. Each column is a new simulation.
}
\description{
Simulate Responses Values Given a Model and New Data
}
