% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdata_perturb_mvhist.R
\name{synthdata_perturb_mvhist}
\alias{synthdata_perturb_mvhist}
\title{Make Approximate DP synthetic data from mixed data types}
\usage{
synthdata_perturb_mvhist(
  data,
  epsilon,
  delta = 0,
  continuous.vars = NULL,
  num.bin = NULL,
  bin.param = NA,
  add.cont.variation = FALSE,
  treatment.colname = "treatment",
  with.treatment = TRUE,
  assign.type = "simple",
  return.time = FALSE,
  rseed = NA,
  within.blocks = TRUE,
  blocks = NULL,
  block.sizes = NULL,
  ...
)
}
\arguments{
\item{data}{a data.frame (or vector of data) to generate (\code{epsilon}, \code{delta})-DP version of}

\item{epsilon}{a positive value that defines the privacy cost}

\item{delta}{a nonnegative value that defines the privacy cost. Default is \code{delta=0}.}

\item{continuous.vars}{vector of strings, indices, or logical values to indicate
the columns of \code{data} with continuous variables. If \code{continuous.vars=NULL},
all variables are assumed to be categorical.}

\item{num.bin}{a numeric value or vector of numeric values for the number of
bins for each continuous variable. If \code{num.bin=NA}, the \code{bin.param}
is used to determine the number of bins.}

\item{bin.param}{a numeric value in (0,1) to determine number of bins for
continuous variables to be used if \code{num.bins=NA}.}

\item{add.cont.variation}{a logical if \code{TRUE}, add uniform noise to midpoint
to get a random value within the range of bin. Default is \code{FALSE}.}

\item{treatment.colname}{a string to rename the treatment column. Default is "treatment".}

\item{with.treatment}{a logical if \code{TRUE}, randomly assignment treatments
are added using \code{treatment_assign} function. Default is \code{TRUE}.}

\item{assign.type}{a string taking values "simple","complete","block",
"cluster", or "block_and_cluster" to specify what kind of random assignment
should be used. Default is "simple".}

\item{return.time}{a logical to indicate if the computation times should be
returned with the synthetic data. Default is \code{FALSE}.}

\item{rseed}{a value to set the randomization seed. If \code{rseed=NA} (default),
then no random seed it set.}

\item{within.blocks}{a logical to indicate if the data should be sampled within the
block levels in order to maintain the same block sizes as the original data.}

\item{blocks}{column name(s) for the variables for block treatment assignment. This
is only needed if \code{assign.type=="block"} or \code{=="block_and_cluster"}.}

\item{block.sizes}{sample size of each block level. If NULL, then the size of the
blocks in the \code{data} will be used.}

\item{...}{additional arguments are such as \code{blocks} or \code{clusters}
are supplied to the \code{treatment_assign} function.}
}
\value{
a synthetic data.frame for confidential data (inputted as \code{data})
that satisfies (\code{epsilon},\code{delta})-DP using the perturbed
multivariate histogram method.
}
\description{
The synthetic data is created using perturbed multivariate
histogram methods  that satisfy (\code{epsilon}, \code{delta})-differential privacy. The
methods are introduced by Dwork et al. (2006) and Wasserman and Zhou (2008).
}
\examples{
synthdata_perturb_mvhist(mtcars[,2:6],1,continuous.vars=c(FALSE,TRUE,TRUE,TRUE,TRUE),num.bin=3)

}
\references{
Dwork, C., McSherry, F., Nissim, K., and Smith, A. (2006). Calibrating noise to
sensitivity in private data analysis. volume Vol. 3876, pages 265–284.
Wasserman, L. and Zhou, S. (2008). A statistical framework for differential privacy.
}
\seealso{
Other syntheticData: 
\code{\link{dp_perturbed_hist}()},
\code{\link{multivariate_histogram}()},
\code{\link{treatment_assign}()}
}
\concept{syntheticData}
