% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_assign.R
\name{treatment_assign}
\alias{treatment_assign}
\title{Assign Treatments Using randomizr Package Functions}
\usage{
treatment_assign(
  synth.data,
  assign.type,
  num.arms = NULL,
  treatment.colname = "treatment",
  blocks = NULL,
  clusters = NULL,
  ...
)
}
\arguments{
\item{synth.data}{a data.frame with the covariates including any blocking or
clustering variables to be used in assignment}

\item{assign.type}{a string taking values "simple","complete","block","cluster",
or "block_and_cluster" to specify what kind of random assignment should be used.
See Details.}

\item{num.arms}{is the number of treatment arms. If \code{num.arms=NULL}, the value
may be determined by additional arguments passed to \code{treatment_assignment}.
If not determined by these arguments, then there will be two treatment conditions.}

\item{treatment.colname}{a string to rename the treatment column. Default is "treatment"}

\item{blocks}{is a vector of indices, column names, or logical to specify
what columns of \code{synth.data} to use as blocking variables in random
treatment assignment. If \code{assign.treatment} takes value "block" or
"block_and_cluster" a value for \code{blocks} must be supplied.}

\item{clusters}{is a vector of indices, column names, or logical to specify
what columns of \code{synth.data} to use as clustering variables in random
treatment assignment. If \code{assign.treatment} takes value "cluster" or
"block_and_cluster" a value for \code{clusters} must be supplied.}

\item{...}{Additional arguments will be passed to the corresponding \pkg{randomizr}
function, where the function is determined by \code{assign.type}.}
}
\value{
a data.frame that adds a treatment column to synth.data.
}
\description{
Assign Treatments Using randomizr Package Functions
}
\details{
Additional inputs in the function are supplied to the respective \pkg{randomizr} function.
The randomized treatment assignment come from functions in \pkg{randomizr}.
If \code{assign.type}="simple", \code{\link{simple_ra}} is used.
If \code{assign.type}="complete", \code{\link{complete_ra}} is used.
If \code{assign.type}="block", \code{\link{block_ra}}
is used and \code{blocks} must be supplied in the input.
If \code{assign.type}="cluster", \code{\link{cluster_ra}}
is used and \code{clusters} must be supplied in the input.
If \code{assign.type}="block_and_cluster",
\code{\link{block_and_cluster_ra}}
is used and \code{blocks} and \code{clusters} must be supplied in the input.
}
\examples{
set.seed(1)
data=data.frame("block.cov"=rep(c("A","A","B","B","B","C","C"),20),
                "covariate"=rnorm(140,48,3))
treatment_assign(data,assign.type="block",num.arms=3,blocks="block.cov")

}
\seealso{
Other syntheticData: 
\code{\link{dp_perturbed_hist}()},
\code{\link{multivariate_histogram}()},
\code{\link{synthdata_perturb_mvhist}()}
}
\concept{syntheticData}
